fix: dolist module/class mismatch
