+++ app/lib/dolist/ignorable_error.rb
@@ -1,0 +1,3 @@
+module Dolist
+  class IgnorableError < StandardError; end
+end
